rename modules to reducers
