+++ src/containers/app/index.js
@@ -6,7 +6,6 @@ const App = () => (
   <div>
     <header>
       <Link to="/">Home</Link>
-      <Link to="/about">About</Link>
     </header>
 
     <main>
