+++ src/containers/home/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { fetchGreeting } from '../../modules/greeting';
+import { fetchGreeting } from '../../reducers/greeting';
 
 const Home = props => (
   <div>
